requirements.txt
streamlit-authenticator==0.2.1
scipy==1.8.1
seaborn==0.12.1
streamlit==1.14.0
matplotlib==3.5.2
pyyaml==6.0
tqdm==4.64.1
numpy==1.23.4